test(pagination): fix test when two Next page buttons are rendered
'Next page' aria-label was missing from the Button with chevron because aria-label is not passed to <a> JSX element in Button component. Now that we've replaced Button with <a>, the aria-label is present and so we have multiple elements with 'button' role and 'Next page' label when Pagination is in first page view.